Refactored the whole uuid command #4
