--- cmd/uuid/v5.go
@@ -1,67 +1,0 @@
-// Copyright © 2018 Christian Claus <ch.claus@me.com>
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-
-package uuid
-
-import (
-	"fmt"
-	"github.com/chclaus/dt/config"
-	"github.com/pkg/errors"
-	"github.com/satori/go.uuid"
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-)
-
-// uuidV5Cmd represents the uuidV5 command
-var uuidV5Cmd = &cobra.Command{
-	Use:   "v5",
-	Short: "Generates a UUID Version 5 (namespace(UUID), value)",
-	Long:  "Generates a v5 UUID, based on SHA1 hashing of (namespace(UUID), value) (RFC 4122)",
-	PreRunE: func(cmd *cobra.Command, args []string) error {
-		ns := config.Cfg.UUID.Namespace
-		errMsg := "Error: You have to specify a namespace and value. The namespace MUST be a valid UUID"
-		if ns == "" {
-			return errors.New(errMsg)
-		}
-		if _, err := uuid.FromString(ns); err != nil {
-			return errors.New(errMsg)
-		}
-
-		return nil
-	},
-	Run: func(cmd *cobra.Command, args []string) {
-		ns := cmd.Flag("namespace").Value.String()
-		nsUUID, _ := uuid.FromString(ns)
-		val := cmd.Flag("value").Value.String()
-
-		fmt.Println(uuid.NewV5(nsUUID, val).String())
-	},
-	Example: "dt uuid v5 -n cacae610-c76a-4736-90ef-0271126b4345 -v foo",
-}
-
-func init() {
-	uuidCmd.AddCommand(uuidV5Cmd)
-
-	uuidV5Cmd.Flags().StringP("namespace", "n", "", "The namespace that should be hashed. It should be a domain or application specific UUID")
-	uuidV5Cmd.Flags().StringP("value", "v", "", "The value that should be hashed")
-	uuidV5Cmd.MarkFlagRequired("value")
-
-	viper.BindPFlag("uuid.namespace", uuidV5Cmd.Flags().Lookup("namespace"))
-}
