+++ cmd/version/version.go
@@ -32,9 +32,8 @@ var versionCmd = &cobra.Command{
 	Short: "Prints the current version of the dt",
 	Long:  "All software has versions. This is dt's",
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("dt - the dev toolbelt v0.2.0 🤓")
+		fmt.Println("dt - the dev toolbelt v1.0.0 🤓")
 	},
-	Example: ``,
 }
 
 func init() {
